% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamfluence-package.R
\docType{package}
\name{gamfluence-package}
\alias{gamfluence}
\alias{gamfluence-package}
\title{gamfluence: deletion influence diagnostics for penalised GAMs}
\description{
Site (cluster) deletion diagnostics for models fitted with \code{\link[mgcv:gam]{mgcv::gam()}}:
each site is removed and the model refitted with the estimated
hyperparameters fixed and re-estimated, on the frozen full-data
representation. The main function is \code{\link[=gam_influence]{gam_influence()}}.
}
\author{
\strong{Maintainer}: Ziqi Song \email{ziqisong030@gmail.com}

}
\keyword{internal}
